backend/requirements.txt
Flask==2.2.5
flask-cors==4.0.0
mysql-connector-python==8.1.0
